requirements.txt
streamlit>=1.29.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=2.2.4
python-dotenv>=1.0.0
torch
endee
